requirements.txt
fastapi[all]
uvicorn[standard]
langchain
langchain-groq
openai
python-dotenv
